+++ tests/test_webapp.py
@@ -1,5 +1,9 @@
 from tests.job_store_cases import JobStoreTests
+from tests.movie_export_cases import MovieExportTests
 from tests.studio_manager_cases import StudioManagerTests
 from tests.webapp_route_cases import AppFactoryTests, LauncherScriptTests
 
-__all__ = ["AppFactoryTests", "JobStoreTests", "LauncherScriptTests", "StudioManagerTests"]
+__all__ = [
+    "AppFactoryTests", "JobStoreTests", "LauncherScriptTests",
+    "MovieExportTests", "StudioManagerTests",
+]